# src/main/resources/static/custom-swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"Custom swagger",
    "version":"1.0",
    "title":"Swagger UI",
    "contact":{
      "name":"mac carthy",
      "url":"http://mac.carthy.org",
      "email":"mac.carthy@ctu.org"
    },
    "license":{
      "name":"apache",
      "url":"http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host":"localhost:8080",
  "basePath":"/",
  "tags":[
    {
      "name":"user-controller",
      "description":"User Controller"
    }
  ],
  "schemes":[
    "http"
  ],
  "paths":{
    "/v1.0/users":{
      "post":{
        "tags":[
          "user-controller"
        ],
        "summary":"addUser",
        "operationId":"addUserUsingPOST",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json;charset=UTF-8"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"user",
            "description":"user",
            "required":true,
            "schema":{
              "$ref":"#/definitions/User"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/User"
            }
          },
          "401":{
            "description":"Unauthorized"
          }
        },
        "deprecated":false
      }
    },
    "/v1.0/users/{userId}":{
      "get":{
        "tags":[
          "user-controller"
        ],
        "summary":"/{userId}",
        "operationId":"getUserByIdUsingGET",
        "produces":[
          "application/json;charset=UTF-8"
        ],
        "parameters":[
          {
            "name":"userId",
            "in":"path",
            "description":"userId",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/User"
            }
          },
          "204":{
            "description":"No content",
            "schema":{
              "$ref":"#/definitions/User"
            }
          },
          "401":{
            "description":"Unauthorized"
          }
        },
        "deprecated":false
      }
    }
  },
  "securityDefinitions":{
    "basicAuth":{
      "type":"basic"
    }
  },
  "definitions":{
    "Address":{
      "type":"object",
      "properties":{
        "number":{
          "type":"integer",
          "format":"int32",
          "minimum":1.0,
          "maximum":100.0,
          "exclusiveMinimum":false,
          "exclusiveMaximum":false
        },
        "postalCode":{
          "type":"string",
          "pattern":"[0-9]{5}"
        },
        "streetName":{
          "type":"string",
          "minLength":5,
          "maxLength":255
        }
      },
      "title":"Address"
    },
    "ContactMedium":{
      "type":"object",
      "required":[
        "type"
      ],
      "properties":{
        "type":{
          "type":"string",
          "enum":[
            "TELEPHONE",
            "EMAIL"
          ]
        }
      },
      "title":"ContactMedium"
    },
    "Email":{
      "title":"Email",
      "allOf":[
        {
          "$ref":"#/definitions/ContactMedium"
        },
        {
          "type":"object",
          "required":[
            "type"
          ],
          "properties":{
            "email":{
              "type":"string"
            },
            "type":{
              "type":"string",
              "enum":[
                "TELEPHONE",
                "EMAIL"
              ]
            }
          },
          "title":"Email"
        }
      ]
    },
    "Telephone":{
      "title":"Telephone",
      "allOf":[
        {
          "$ref":"#/definitions/ContactMedium"
        },
        {
          "type":"object",
          "required":[
            "type"
          ],
          "properties":{
            "number":{
              "type":"string"
            },
            "type":{
              "type":"string",
              "enum":[
                "TELEPHONE",
                "EMAIL"
              ]
            }
          },
          "title":"Telephone"
        }
      ]
    },
    "User":{
      "type":"object",
      "required":[
        "address"
      ],
      "properties":{
        "address":{
          "$ref":"#/definitions/Address"
        },
        "email":{
          "$ref":"#/definitions/Email"
        },
        "fullName":{
          "type":"string"
        },
        "telephone":{
          "$ref":"#/definitions/Telephone"
        },
        "userId":{
          "type":"integer",
          "format":"int64"
        }
      },
      "title":"User"
    }
  }
}